==> src/3-aim2/2-analysis/0-run-scripts.R <==

# source(paste0(here::here(),"/src/3-aim2/1-data cleaning/merge_WBB_ch_env.R"))
# source(paste0(here::here(),"/src/3-aim2/1-data cleaning/merge_mapsan_ch_env.R"))

source(paste0(here::here(),"/src/3-aim2/2-analysis/1-combine-health-datasets.R"))
source(paste0(here::here(),"/src/3-aim2/2-analysis/2-aim-2-clean-covariates.R"))

source(paste0(here::here(),"/src/3-aim2/2-analysis/3-aim-2-analysis.R"))
source(paste0(here::here(),"/src/3-aim2/2-analysis/3-aim-2-abundance-analysis.R"))
source(paste0(here::here(),"/src/3-aim2/2-analysis/3-aim-2-analysis-all-diarrhea.R"))
source(paste0(here::here(),"/src/3-aim2/2-analysis/3-aim-2-pathogen-specific-analysis.R"))
source(paste0(here::here(),"/src/3-aim2/2-analysis/4-aim-2-subgroup-analysis.R"))
source(paste0(here::here(),"/src/3-aim2/2-analysis/4-aim-2-subgroup-analysis-age.R"))
source(paste0(here::here(),"/src/3-aim2/2-analysis/5-pool-results-aim2.R"))
#source(paste0(here::here(),"/src/3-aim2/2-analysis/3-aim-2-analysis-tmle2.R"))
source(paste0(here::here(),"/src/3-aim2/2-analysis/6-compile-tmle.R"))
source(paste0(here::here(),"/src/3-aim2/2-analysis/7-aim-2-tr-analysis.R"))
